refactor: VNC connection creation to own module
