+++ client/src/vnc/mod.rs
@@ -1,0 +1,4 @@
+mod connect;
+mod auth;
+
+pub use self::connect::connect;
